+++ src/Context.php
@@ -12,6 +12,8 @@ namespace DecodeLabs\Fabric;
 use DecodeLabs\Clip\Controller as ClipController;
 use DecodeLabs\Fabric;
 use DecodeLabs\Genesis\Context as Genesis;
+use DecodeLabs\Systemic;
+use DecodeLabs\Terminus as Cli;
 use DecodeLabs\Veneer;
 use DecodeLabs\Veneer\LazyLoad;
 
@@ -40,6 +42,26 @@ class Context
     public function getTaskController(): ClipController
     {
         return $this->genesis->container->get(ClipController::class);
+    }
+
+    /**
+     * Ensure CLI is running in source mode
+     */
+    public function ensureCliSource(): void
+    {
+        if (!$this->genesis->build->isCompiled()) {
+            return;
+        }
+
+        Cli::notice('Switching to source mode');
+        Cli::newLine();
+
+        $args = $_SERVER['argv'] ?? [];
+        $path = realpath(array_shift($args));
+        $args[] = '--fabric-source';
+
+        Systemic::runScript([$path, ...$args]);
+        $this->genesis->shutdown();
     }
 }